+++ common/src/main/java/net/mehvahdjukaar/jeed/plugin/rei/ingredient/EffectInstanceRenderer.java
@@ -73,6 +73,6 @@ public class EffectInstanceRenderer extends EffectRenderer implements BatchedEnt
 
     @Override
     public @Nullable Tooltip getTooltip(EntryStack<MobEffectInstance> entry, TooltipContext context) {
-        return Tooltip.create(getTooltipsWithDescription(entry.getValue(), context.getFlag(), false));
+        return Tooltip.create(getTooltipsWithDescription(entry.getValue(), context.getFlag(), false, false));
     }
 }