+++ forge/src/main/java/net/mehvahdjukaar/jeed/forge/JeedClient.java
@@ -3,7 +3,7 @@ package net.mehvahdjukaar.jeed.forge;
 import net.mehvahdjukaar.jeed.Jeed;
 import net.mehvahdjukaar.jeed.api.IEffectScreenExtension;
 import net.mehvahdjukaar.jeed.common.ScreenExtensionsHandler;
-import net.mehvahdjukaar.jeed.compat.forge.NativeCompat;
+import net.mehvahdjukaar.jeed.compat.NativeCompat;
 import net.mehvahdjukaar.jeed.compat.forge.StylishEffectsCompat;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraftforge.client.event.ScreenEvent;
@@ -43,7 +43,8 @@ public class JeedClient {
                     } else {
                         var effect = ext.getEffectAtPosition(screen, event.getMouseX(), event.getMouseY(), IEffectScreenExtension.CallReason.TOOLTIP);
                         if (effect != null) {
-                            ScreenExtensionsHandler.renderEffectTooltip(effect, screen, event.getPoseStack(), event.getMouseX(), event.getMouseY());
+                            ScreenExtensionsHandler.renderEffectTooltip(effect, screen, event.getPoseStack(),
+                                    event.getMouseX(), event.getMouseY(), ext.showDurationOnTooltip());
                         }
                     }
                 }